Refactor JournalTable and extract common functionality into base class.
