+++ webapp/src/components/journal-table.scss
@@ -1,22 +1,8 @@
 
 
-.select {
-  margin-left: 20px;
-  margin-right: 20px;
-  display: inline-block;
-  width: auto;
-}
-
-
-
-.filter {
-  margin-bottom: 20px;
-}
-
 .heading {
   cursor: pointer;
 }
-
 
 .logo {
   height: 16px;
