+++ webapp/src/components/table.scss
@@ -1,0 +1,10 @@
+.select {
+  margin-left: 20px;
+  margin-right: 20px;
+  display: inline-block;
+  width: auto;
+}
+
+.filter {
+  margin-bottom: 20px;
+}
